move functions to utils and test_utils
